test(core): split large app builder and watch tests
